fix bugs after run tests

+++ src/db_drivers/vector_driver/connectors/ChromaConnector.py
@@ -51,9 +51,11 @@ class ChromaConnection(AbstractVectorDatabaseConnection):
                 ids=list(map(lambda idx: items[idx].id, insts_wo_md)))
 
     def read(self, ids: List[str], includes: List[str] = ['embeddings', 'documents'], **kwargs) -> List[VectorDBInstance]:
-        raw_instances = self.collection.get(
-            include=includes,
-            ids=ids, **kwargs)
+        raw_instances = []
+        if len(ids):
+            raw_instances = self.collection.get(
+                include=includes,
+                ids=ids, **kwargs)
 
         formates_instances = []
         for i in range(len(raw_instances['ids'])):
@@ -68,7 +70,8 @@ class ChromaConnection(AbstractVectorDatabaseConnection):
         pass
 
     def delete(self, ids: List[str], **kwargs) -> None:
-        self.collection.delete(ids=ids, **kwargs)
+        if len(ids):
+            self.collection.delete(ids=ids, **kwargs)
 
     def retrieve(
             self, query_instances: List[VectorDBInstance], n_results: int = 50,
@@ -77,7 +80,7 @@ class ChromaConnection(AbstractVectorDatabaseConnection):
         n_results = collection_size if collection_size < n_results else n_results
 
         raw_retrieved_instances = self.collection.query(
-            query_embeddings=[inst.embedding.tolist() for inst in query_instances],
+            query_embeddings=[inst.embedding for inst in query_instances],
             include=includes + ['distances'], n_results=n_results, **kwargs)
 
         formated_instances = []
@@ -97,8 +100,8 @@ class ChromaConnection(AbstractVectorDatabaseConnection):
         return self.collection.count()
 
     def item_exist(self, id: str) -> bool:
-        # TODO
-        pass
+        output = self.collection.get(ids=[id])
+        return len(output) > 0
 
     def clear(self) -> None:
         self.client.delete_collection(name=self.config.db_name)
